OT148-93: Rename class to NewsControllerFunctionalTest
